+++ src/Submit.jsx
@@ -1,14 +1,35 @@
 import Button from "./components/Button/Button";
 
-
 const Submit = () => {
+    // form kontrol
+    const handleSubmit = (e) => {
+        e.preventDefault();
+
+        const name = e.target.name.value.trim();
+        const email = e.target.email.value.trim();
+        const message = e.target.message.value.trim();
+
+        if (!name || !email || !message) {
+            alert("Lütfen tüm alanları doldurun!");
+            return;
+        }
+
+        alert("Form başarıyla gönderildi!");
+    };
+
+
     return (
         <div className="Submit">
             <h2>İletişim Formu</h2>
 
-            <form>
+            <form onSubmit={handleSubmit}>
                 <label htmlFor="name">Ad Soyad</label>
-                <input type="text" id="name" name="name" placeholder="Adınızı girin" />
+                <input
+                    type="text"
+                    id="name"
+                    name="name"
+                    placeholder="Adınızı girin"
+                />
 
                 <label htmlFor="email">E-posta</label>
                 <input
@@ -27,9 +48,7 @@ const Submit = () => {
                 ></textarea>
 
                 <Button type="submit">Gönder</Button>
-
             </form>
-
         </div>
     );
 };